sprite constraint 1/4
