+++ cmake-build-debug/CMakeFiles/Makefile2
@@ -94,12 +94,12 @@ clean: CMakeFiles/cub3D.dir/clean
 CMakeFiles/cub3D.dir/all:
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/cub3D.dir/build.make CMakeFiles/cub3D.dir/depend
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/cub3D.dir/build.make CMakeFiles/cub3D.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles --progress-num=1,2,3,4 "Built target cub3D"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles --progress-num=1,2,3,4,5 "Built target cub3D"
 .PHONY : CMakeFiles/cub3D.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/cub3D.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles 4
+	$(CMAKE_COMMAND) -E cmake_progress_start /Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles 5
 	$(MAKE) $(MAKESILENT) -f CMakeFiles/Makefile2 CMakeFiles/cub3D.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start /Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles 0
 .PHONY : CMakeFiles/cub3D.dir/rule
