+++ cmake-build-debug/CMakeFiles/cub3D.dir/build.make
@@ -116,11 +116,25 @@ CMakeFiles/cub3D.dir/main.c.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling C source to assembly CMakeFiles/cub3D.dir/main.c.s"
 	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -S /Users/cveeta/CLionProjects/cub3D/main.c -o CMakeFiles/cub3D.dir/main.c.s
 
+CMakeFiles/cub3D.dir/sprt.c.o: CMakeFiles/cub3D.dir/flags.make
+CMakeFiles/cub3D.dir/sprt.c.o: ../sprt.c
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_4) "Building C object CMakeFiles/cub3D.dir/sprt.c.o"
+	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -o CMakeFiles/cub3D.dir/sprt.c.o   -c /Users/cveeta/CLionProjects/cub3D/sprt.c
+
+CMakeFiles/cub3D.dir/sprt.c.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing C source to CMakeFiles/cub3D.dir/sprt.c.i"
+	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -E /Users/cveeta/CLionProjects/cub3D/sprt.c > CMakeFiles/cub3D.dir/sprt.c.i
+
+CMakeFiles/cub3D.dir/sprt.c.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling C source to assembly CMakeFiles/cub3D.dir/sprt.c.s"
+	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -S /Users/cveeta/CLionProjects/cub3D/sprt.c -o CMakeFiles/cub3D.dir/sprt.c.s
+
 # Object files for target cub3D
 cub3D_OBJECTS = \
 "CMakeFiles/cub3D.dir/gnl/get_next_line.c.o" \
 "CMakeFiles/cub3D.dir/gnl/get_next_line_utils.c.o" \
-"CMakeFiles/cub3D.dir/main.c.o"
+"CMakeFiles/cub3D.dir/main.c.o" \
+"CMakeFiles/cub3D.dir/sprt.c.o"
 
 # External object files for target cub3D
 cub3D_EXTERNAL_OBJECTS =
@@ -128,9 +142,10 @@ cub3D_EXTERNAL_OBJECTS =
 cub3D: CMakeFiles/cub3D.dir/gnl/get_next_line.c.o
 cub3D: CMakeFiles/cub3D.dir/gnl/get_next_line_utils.c.o
 cub3D: CMakeFiles/cub3D.dir/main.c.o
+cub3D: CMakeFiles/cub3D.dir/sprt.c.o
 cub3D: CMakeFiles/cub3D.dir/build.make
 cub3D: CMakeFiles/cub3D.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_4) "Linking C executable cub3D"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/Users/cveeta/CLionProjects/cub3D/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_5) "Linking C executable cub3D"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/cub3D.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
