+++ cmake-build-debug/CMakeFiles/cub3D.dir/link.txt
@@ -1,1 +1,1 @@
-/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc -lmlx -framework OpenGL -framework AppKit -g -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.15.sdk -mmacosx-version-min=10.14 -Wl,-search_paths_first -Wl,-headerpad_max_install_names  CMakeFiles/cub3D.dir/gnl/get_next_line.c.o CMakeFiles/cub3D.dir/gnl/get_next_line_utils.c.o CMakeFiles/cub3D.dir/main.c.o  -o cub3D   -L/Users/cveeta/CLionProjects/cub3D/. 
+/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc -lmlx -framework OpenGL -framework AppKit -g -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.15.sdk -mmacosx-version-min=10.14 -Wl,-search_paths_first -Wl,-headerpad_max_install_names  CMakeFiles/cub3D.dir/gnl/get_next_line.c.o CMakeFiles/cub3D.dir/gnl/get_next_line_utils.c.o CMakeFiles/cub3D.dir/main.c.o CMakeFiles/cub3D.dir/sprt.c.o  -o cub3D   -L/Users/cveeta/CLionProjects/cub3D/.
