+++ cmake-build-debug/CMakeFiles/cub3D.dir/progress.make
@@ -2,4 +2,5 @@ CMAKE_PROGRESS_1 = 1
 CMAKE_PROGRESS_2 = 2
 CMAKE_PROGRESS_3 = 3
 CMAKE_PROGRESS_4 = 4
+CMAKE_PROGRESS_5 = 5
