+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Feb 07 19:24 MSK
+Start testing: Feb 08 22:27 MSK
 ----------------------------------------------------------
-End testing: Feb 07 19:24 MSK
+End testing: Feb 08 22:27 MSK
